requirements.txt
pyyaml
argparse
rich
colorama